backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.9
pydantic==1.10.14
alembic==1.11.3
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.24.1
